+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/file/FlexibleSearchFileToolbarInstaller.kt
@@ -20,14 +20,20 @@ package com.intellij.idea.plugin.hybris.flexibleSearch.file
 
 import com.intellij.idea.plugin.hybris.startup.event.AbstractHybrisFileToolbarInstaller
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.editor.ex.EditorEx
+import com.intellij.openapi.fileTypes.FileType
+import com.intellij.openapi.project.Project
 
 class FlexibleSearchFileToolbarInstaller : AbstractHybrisFileToolbarInstaller(
     "hybris.fxs.console",
     "hybris.fxs.toolbar.left",
-    "hybris.fxs.toolbar.right"
+    "hybris.fxs.toolbar.right",
+    FlexibleSearchFileType.instance
 ) {
 
     companion object {
         val instance: FlexibleSearchFileToolbarInstaller = ApplicationManager.getApplication().getService(FlexibleSearchFileToolbarInstaller::class.java)
     }
+
+    override fun isToolbarEnabled(project: Project, editor: EditorEx) = true
 }